# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API sederhana untuk mengelola koleksi musik",
        "title": "Music API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/musics": {
            "get": {
                "description": "Ambil semua data musik",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "musics"
                ],
                "summary": "Get all musics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Music"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Tambahkan musik ke dalam koleksi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "musics"
                ],
                "summary": "Tambah musik baru",
                "parameters": [
                    {
                        "description": "Music Data",
                        "name": "music",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Music"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Music"
                        }
                    }
                }
            }
        },
        "/musics/{id}": {
            "get": {
                "description": "Ambil data musik berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "musics"
                ],
                "summary": "Get music by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Music ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Music"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit data musik berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "musics"
                ],
                "summary": "Update musik",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Music ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Music Data",
                        "name": "music",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Music"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Music"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Hapus data musik berdasarkan ID",
                "tags": [
                    "musics"
                ],
                "summary": "Hapus musik",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Music ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Music": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string",
                    "example": "A Night at the Opera"
                },
                "artist": {
                    "type": "string",
                    "example": "Queen"
                },
                "genre": {
                    "type": "string",
                    "example": "Rock"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "length": {
                    "type": "integer",
                    "example": 5
                },
                "name": {
                    "type": "string",
                    "example": "Bohemian Rhapsody"
                },
                "writer": {
                    "type": "string",
                    "example": "Freddie Mercury"
                },
                "year": {
                    "type": "integer",
                    "example": 1975
                }
            }
        }
    }
}